Create proper Flask app
