+++ .gitignore
@@ -84,4 +84,4 @@ Temporary Items
 .apdisk
 
 *.db
-morseweb/web/models/*
+morseweb/static/json/*
