+++ morseweb/static/js/EditorControls.js
