--- morseweb/webapp/static/json/empty.json
@@ -1,84 +1,0 @@
-{
-    "metadata": {
-        "version": 4.4,
-        "type": "Object",
-        "generator": "io_three",
-        "sourceFile": "empty.blend"
-    },
-    "materials": [{
-        "depthTest": true,
-        "depthWrite": true,
-        "ambient": 10724259,
-        "color": 10724259,
-        "emissive": 0,
-        "shininess": 50,
-        "vertexColors": false,
-        "specular": 8355711,
-        "blending": "NormalBlending",
-        "name": "Material",
-        "type": "MeshPhongMaterial",
-        "uuid": "FC9C3C1C-87E7-3748-8148-105003E5F3D5"
-    }],
-    "geometries": [{
-        "materials": [{
-            "wireframe": false,
-            "colorDiffuse": [0.64,0.64,0.64],
-            "depthTest": true,
-            "depthWrite": true,
-            "visible": true,
-            "blending": "NormalBlending",
-            "transparent": false,
-            "DbgName": "Material",
-            "colorSpecular": [0.5,0.5,0.5],
-            "DbgColor": 15658734,
-            "colorAmbient": [0.64,0.64,0.64],
-            "doubleSided": true,
-            "specularCoef": 50,
-            "DbgIndex": 0,
-            "colorEmissive": [0,0,0],
-            "opacity": 1,
-            "shading": "phong"
-        }],
-        "type": "Geometry",
-        "uuid": "FE46F7EB-A263-3807-8987-EB40ABE18EAA",
-        "data": {
-            "normals": [0.577349,0.577349,-0.577349,0.577349,-0.577349,-0.577349,-0.577349,-0.577349,-0.577349,-0.577349,0.577349,-0.577349,0.577349,0.577349,0.577349,-0.577349,0.577349,0.577349,-0.577349,-0.577349,0.577349,0.577349,-0.577349,0.577349],
-            "metadata": {
-                "materials": 1,
-                "version": 3,
-                "normals": 8,
-                "uvs": 0,
-                "vertices": 8,
-                "faces": 6,
-                "generator": "io_three"
-            },
-            "name": "CubeGeometry",
-            "uvs": [],
-            "vertices": [1,1,-1,1,-1,-1,-1,-1,-1,-1,1,-1,1,0.999999,1,0.999999,-1,1,-1,-1,1,-1,1,1],
-            "faces": [35,0,1,2,3,0,0,1,2,3,35,4,7,6,5,0,4,5,6,7,35,0,4,5,1,0,0,4,7,1,35,1,5,6,2,0,1,7,6,2,35,2,6,7,3,0,2,6,5,3,35,4,0,3,7,0,4,0,3,5]
-        }
-    }],
-    "animations": [{
-        "name": "default",
-        "tracks": [],
-        "fps": 24
-    }],
-    "images": [],
-    "textures": [],
-    "object": {
-        "matrix": [1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1],
-        "type": "Scene",
-        "uuid": "581759EC-4EC5-400E-8267-8372411334E2",
-        "children": [{
-            "name": "Ground",
-            "uuid": "29A30250-63C3-3FC2-9E32-2D5909D17197",
-            "matrix": [-20,0,0,0,0,0,20,0,0,0.065215,0,0,0,-0.06529,0,1],
-            "visible": true,
-            "type": "Mesh",
-            "material": "FC9C3C1C-87E7-3748-8148-105003E5F3D5",
-            "castShadow": true,
-            "receiveShadow": true,
-            "geometry": "FE46F7EB-A263-3807-8987-EB40ABE18EAA"
-        }]
-    }
-}
