--- morseweb/webapp/static/json/epuck.json
@@ -1,83 +1,0 @@
-{
-    "images": [],
-    "materials": [{
-        "emissive": 0,
-        "blending": "NormalBlending",
-        "name": "Material.004",
-        "shininess": 50,
-        "specular": 8355711,
-        "uuid": "52D21D59-A717-391E-82F2-F87CBF4CE829",
-        "ambient": 2963363,
-        "depthWrite": true,
-        "type": "MeshPhongMaterial",
-        "color": 2963363,
-        "vertexColors": false,
-        "depthTest": true
-    }],
-    "animations": [{
-        "tracks": [],
-        "fps": 24,
-        "name": "default"
-    }],
-    "object": {
-        "type": "Scene",
-        "uuid": "A8847297-9754-4AE3-8965-E5518AC16916",
-        "children": [{
-            "name": "Body",
-            "uuid": "AFD86FED-2C5D-316D-846B-16D9B79CD540",
-            "matrix": [-1,0,0,0,0,0,1,0,0,1,0,0,0,0,0,1],
-            "visible": true,
-            "type": "Mesh",
-            "material": "52D21D59-A717-391E-82F2-F87CBF4CE829",
-            "castShadow": true,
-            "receiveShadow": true,
-            "geometry": "65010417-33DE-3CC1-8750-CBCD89AB3E27"
-        }],
-        "matrix": [1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1]
-    },
-    "metadata": {
-        "sourceFile": "epuck.blend",
-        "version": 4.4,
-        "type": "Object",
-        "generator": "io_three"
-    },
-    "textures": [],
-    "geometries": [{
-        "type": "Geometry",
-        "uuid": "65010417-33DE-3CC1-8750-CBCD89AB3E27",
-        "materials": [{
-            "transparent": false,
-            "colorAmbient": [0.17868,0.219016,0.64],
-            "colorSpecular": [0.5,0.5,0.5],
-            "blending": "NormalBlending",
-            "opacity": 1,
-            "specularCoef": 50,
-            "shading": "phong",
-            "DbgIndex": 0,
-            "visible": true,
-            "DbgColor": 15658734,
-            "colorEmissive": [0,0,0],
-            "wireframe": false,
-            "depthWrite": true,
-            "colorDiffuse": [0.17868,0.219016,0.64],
-            "depthTest": true,
-            "DbgName": "Material.004"
-        }],
-        "data": {
-            "normals": [0,0.727866,-0.68569,0,0.727866,0.68569,0.142003,0.713889,0.68569,0.142003,0.713889,-0.68569,0.278542,0.672475,0.68569,0.278542,0.672475,-0.68569,0.40437,0.605213,0.68569,0.40437,0.605213,-0.68569,0.514664,0.514664,0.68569,0.514664,0.514664,-0.68569,0.605213,0.40437,0.68569,0.605213,0.40437,-0.68569,0.672475,0.278542,0.68569,0.672475,0.278542,-0.68569,0.713889,0.142003,0.68569,0.713889,0.142003,-0.68569,0.727866,0,0.68569,0.727866,0,-0.68569,0.713889,-0.142003,0.68569,0.713889,-0.142003,-0.68569,0.672475,-0.278542,0.68569,0.672475,-0.278542,-0.68569,0.605213,-0.40437,0.68569,0.605213,-0.40437,-0.68569,0.514664,-0.514664,0.68569,0.514664,-0.514664,-0.68569,0.40437,-0.605213,0.68569,0.40437,-0.605213,-0.68569,0.278542,-0.672475,0.68569,0.278542,-0.672475,-0.68569,0.142003,-0.713889,0.68569,0.142003,-0.713889,-0.68569,0,-0.727866,0.68569,0,-0.727866,-0.68569,-0.142003,-0.713889,0.68569,-0.142003,-0.713889,-0.68569,-0.278542,-0.672475,0.68569,-0.278542,-0.672475,-0.68569,-0.40437,-0.605213,0.68569,-0.40437,-0.605213,-0.68569,-0.514695,-0.514664,0.68569,-0.514695,-0.514664,-0.68569,-0.605213,-0.40437,0.68569,-0.605213,-0.40437,-0.68569,-0.672475,-0.278542,0.68569,-0.672475,-0.278542,-0.68569,-0.713889,-0.142003,0.68569,-0.713889,-0.142003,-0.68569,-0.727866,0,0.68569,-0.727866,0,-0.68569,-0.713889,0.142003,0.68569,-0.713889,0.142003,-0.68569,-0.672475,0.278542,0.68569,-0.672475,0.278542,-0.68569,-0.605213,0.40437,0.68569,-0.605213,0.40437,-0.68569,-0.514664,0.514695,0.68569,-0.514664,0.514695,-0.68569,-0.40437,0.605213,0.68569,-0.40437,0.605213,-0.68569,-0.278542,0.672475,0.68569,-0.278542,0.672475,-0.68569,-0.142003,0.713889,0.68569,-0.142003,0.713889,-0.68569],
-            "vertices": [0,0.275,-0.15,0,0.275,0.15,0.05365,0.269716,-0.15,0.05365,0.269716,0.15,0.105238,0.254067,-0.15,0.105238,0.254067,0.15,0.152782,0.228654,-0.15,0.152782,0.228654,0.15,0.194454,0.194454,-0.15,0.194454,0.194454,0.15,0.228654,0.152782,-0.15,0.228654,0.152782,0.15,0.254067,0.105238,-0.15,0.254067,0.105238,0.15,0.269716,0.05365,-0.15,0.269716,0.05365,0.15,0.275,0,-0.15,0.275,0,0.15,0.269716,-0.05365,-0.15,0.269716,-0.05365,0.15,0.254067,-0.105238,-0.15,0.254067,-0.105238,0.15,0.228654,-0.152782,-0.15,0.228654,-0.152782,0.15,0.194454,-0.194454,-0.15,0.194454,-0.194454,0.15,0.152782,-0.228654,-0.15,0.152782,-0.228654,0.15,0.105238,-0.254067,-0.15,0.105238,-0.254067,0.15,0.05365,-0.269716,-0.15,0.05365,-0.269716,0.15,-0,-0.275,-0.15,-0,-0.275,0.15,-0.05365,-0.269716,-0.15,-0.05365,-0.269716,0.15,-0.105238,-0.254067,-0.15,-0.105238,-0.254067,0.15,-0.152782,-0.228654,-0.15,-0.152782,-0.228654,0.15,-0.194454,-0.194454,-0.15,-0.194454,-0.194454,0.15,-0.228654,-0.152782,-0.15,-0.228654,-0.152782,0.15,-0.254067,-0.105238,-0.15,-0.254067,-0.105238,0.15,-0.269716,-0.05365,-0.15,-0.269716,-0.05365,0.15,-0.275,0,-0.15,-0.275,0,0.15,-0.269716,0.05365,-0.15,-0.269716,0.05365,0.15,-0.254067,0.105238,-0.15,-0.254067,0.105238,0.15,-0.228654,0.152782,-0.15,-0.228654,0.152782,0.15,-0.194454,0.194455,-0.15,-0.194454,0.194455,0.15,-0.152781,0.228654,-0.15,-0.152781,0.228654,0.15,-0.105238,0.254067,-0.15,-0.105238,0.254067,0.15,-0.053649,0.269716,-0.15,-0.053649,0.269716,0.15],
-            "name": "BodyGeometry",
-            "faces": [35,0,1,3,2,0,0,1,2,3,35,2,3,5,4,0,3,2,4,5,35,4,5,7,6,0,5,4,6,7,35,6,7,9,8,0,7,6,8,9,35,8,9,11,10,0,9,8,10,11,35,10,11,13,12,0,11,10,12,13,35,12,13,15,14,0,13,12,14,15,35,14,15,17,16,0,15,14,16,17,35,16,17,19,18,0,17,16,18,19,35,18,19,21,20,0,19,18,20,21,35,20,21,23,22,0,21,20,22,23,35,22,23,25,24,0,23,22,24,25,35,24,25,27,26,0,25,24,26,27,35,26,27,29,28,0,27,26,28,29,35,28,29,31,30,0,29,28,30,31,35,30,31,33,32,0,31,30,32,33,35,32,33,35,34,0,33,32,34,35,35,34,35,37,36,0,35,34,36,37,35,36,37,39,38,0,37,36,38,39,35,38,39,41,40,0,39,38,40,41,35,40,41,43,42,0,41,40,42,43,35,42,43,45,44,0,43,42,44,45,35,44,45,47,46,0,45,44,46,47,35,46,47,49,48,0,47,46,48,49,35,48,49,51,50,0,49,48,50,51,35,50,51,53,52,0,51,50,52,53,35,52,53,55,54,0,53,52,54,55,35,54,55,57,56,0,55,54,56,57,35,56,57,59,58,0,57,56,58,59,35,58,59,61,60,0,59,58,60,61,34,5,3,1,0,4,2,1,34,1,63,61,0,1,62,60,34,61,59,57,0,60,58,56,34,57,55,53,0,56,54,52,34,53,51,49,0,52,50,48,34,49,47,45,0,48,46,44,34,45,43,41,0,44,42,40,34,41,39,37,0,40,38,36,34,37,35,33,0,36,34,32,34,33,31,29,0,32,30,28,34,29,27,25,0,28,26,24,34,25,23,21,0,24,22,20,34,21,19,17,0,20,18,16,34,17,15,13,0,16,14,12,34,13,11,9,0,12,10,8,34,9,7,5,0,8,6,4,34,5,1,61,0,4,1,60,34,61,57,53,0,60,56,52,34,53,49,45,0,52,48,44,34,45,41,37,0,44,40,36,34,37,33,29,0,36,32,28,34,29,25,21,0,28,24,20,34,21,17,13,0,20,16,12,34,13,9,5,0,12,8,4,34,5,61,53,0,4,60,52,34,53,45,37,0,52,44,36,34,37,29,21,0,36,28,20,34,21,13,5,0,20,12,4,34,5,53,37,0,4,52,36,34,37,21,5,0,36,20,4,35,1,0,62,63,0,1,0,63,62,35,60,61,63,62,0,61,60,62,63,34,62,0,2,0,63,0,3,34,2,4,6,0,3,5,7,34,6,8,10,0,7,9,11,34,10,12,14,0,11,13,15,34,14,16,18,0,15,17,19,34,18,20,22,0,19,21,23,34,22,24,26,0,23,25,27,34,26,28,30,0,27,29,31,34,30,32,34,0,31,33,35,34,34,36,38,0,35,37,39,34,38,40,42,0,39,41,43,34,42,44,46,0,43,45,47,34,46,48,50,0,47,49,51,34,50,52,54,0,51,53,55,34,54,56,58,0,55,57,59,34,58,60,62,0,59,61,63,34,62,2,6,0,63,3,7,34,6,10,14,0,7,11,15,34,14,18,22,0,15,19,23,34,22,26,30,0,23,27,31,34,30,34,38,0,31,35,39,34,38,42,46,0,39,43,47,34,46,50,54,0,47,51,55,34,54,58,62,0,55,59,63,34,62,6,14,0,63,7,15,34,14,22,30,0,15,23,31,34,30,38,46,0,31,39,47,34,46,54,62,0,47,55,63,34,62,14,30,0,63,15,31,34,30,46,62,0,31,47,63],
-            "metadata": {
-                "normals": 64,
-                "vertices": 64,
-                "materials": 1,
-                "generator": "io_three",
-                "version": 3,
-                "faces": 92,
-                "uvs": 0
-            },
-            "uvs": []
-        }
-    }]
-}
