--- morseweb/webapp/static/json/floor.json
@@ -1,84 +1,0 @@
-{
-    "images": [],
-    "materials": [{
-        "emissive": 0,
-        "blending": "NormalBlending",
-        "name": "FloorMat",
-        "shininess": 50,
-        "specular": 8355711,
-        "uuid": "1624E5F0-16F6-36C5-933A-267986E26FA5",
-        "ambient": 12831427,
-        "depthWrite": true,
-        "type": "MeshPhongMaterial",
-        "color": 12831427,
-        "vertexColors": false,
-        "depthTest": true
-    }],
-    "animations": [{
-        "tracks": [],
-        "fps": 25,
-        "name": "default"
-    }],
-    "object": {
-        "type": "Scene",
-        "uuid": "E57C6223-3C35-4E2A-8577-CC6A9546C27D",
-        "children": [{
-            "name": "Floor",
-            "uuid": "BC5EA681-B08F-3A26-86A4-DC0BC5C99234",
-            "matrix": [0.000443,0,-9.72398,0,-9.72398,0,-0.000443,0,0,9.72398,0,0,-1.58947,-0.005,-5.71509,1],
-            "visible": true,
-            "type": "Mesh",
-            "material": "1624E5F0-16F6-36C5-933A-267986E26FA5",
-            "castShadow": true,
-            "receiveShadow": true,
-            "geometry": "FDA35051-77AC-3FCF-B642-34C990EA7344"
-        }],
-        "matrix": [1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1]
-    },
-    "metadata": {
-        "sourceFile": "grande_salle.blend",
-        "version": 4.4,
-        "type": "Object",
-        "generator": "io_three"
-    },
-    "textures": [],
-    "geometries": [{
-        "type": "Geometry",
-        "uuid": "FDA35051-77AC-3FCF-B642-34C990EA7344",
-        "materials": [{
-            "transparent": false,
-            "colorAmbient": [0.766838,0.792373,0.766838],
-            "colorSpecular": [0.5,0.5,0.5],
-            "blending": "NormalBlending",
-            "opacity": 1,
-            "specularCoef": 50,
-            "shading": "phong",
-            "DbgIndex": 0,
-            "visible": true,
-            "DbgColor": 15658734,
-            "colorEmissive": [0,0,0],
-            "wireframe": false,
-            "depthWrite": true,
-            "colorDiffuse": [0.766838,0.792373,0.766838],
-            "depthTest": true,
-            "doubleSided": true,
-            "DbgName": "FloorMat"
-        }],
-        "data": {
-            "normals": [0,0,1,0,0,0.999969],
-            "vertices": [-1,1,0,0.542709,-0.820423,0,0.542709,1,0,-1,-0.67,0,0.542709,-0.67,0,-0.62,1,0,-0.62,-0.820423,0,-0.62,-0.67,0],
-            "name": "PlaneGeometry",
-            "faces": [35,5,0,3,7,0,0,1,0,0,35,7,4,2,5,0,0,0,1,0,35,6,1,4,7,0,0,0,0,0],
-            "metadata": {
-                "normals": 2,
-                "vertices": 8,
-                "materials": 1,
-                "generator": "io_three",
-                "version": 3,
-                "faces": 3,
-                "uvs": 0
-            },
-            "uvs": []
-        }
-    }]
-}
